# src/tsoa/swagger.json
{
	"components": {
...
		"schemas": {
...
			"FailureResponse": {
				"properties": {
					"error": {
						"type": "string"
					}
				},
				"required": [
					"error"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"Message": {
				"properties": {
					"messageID": {
						"type": "number",
						"format": "double"
					},
					"senderID": {
						"type": "number",
						"format": "double"
					},
					"receiverID": {
						"type": "number",
						"format": "double"
					},
					"sendtime": {
						"type": "number",
						"format": "double"
					},
					"type": {
						"type": "number",
						"format": "double"
					},
					"subject": {
						"type": "string"
					},
					"body": {
						"type": "string"
					},
					"deleted": {
						"type": "boolean"
					}
				},
				"required": [
					"messageID",
					"senderID",
					"receiverID",
					"sendtime",
					"type",
					"subject",
					"body",
					"deleted"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SendMessageRequest": {
				"properties": {
					"receiverID": {
						"type": "number",
						"format": "double"
					},
					"subject": {
						"type": "string"
					},
					"body": {
						"type": "string"
					}
				},
				"required": [
					"receiverID",
					"subject",
					"body"
				],
				"type": "object",
				"additionalProperties": false
			},
			"DeleteMessageRequest": {
				"properties": {
					"messageID": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"messageID"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"jwt": {
				"type": "apiKey",
				"name": "access-token",
				"in": "header"
			}
		}
	},
	"info": {
		"title": "ugamela api",
		"version": "1.0.0",
		"description": "bla",
		"license": {
			"name": "AGPL-3.0"
		},
		"contact": {
			"email": "api@ugamela.org",
			"name": "the Administrator"
		}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/messages": {
			"get": {
				"operationId": "GetAllMessages",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Message"
									},
									"type": "array"
								}
							}
						}
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/FailureResponse"
								}
							}
						}
					},
					"401": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/FailureResponse"
								}
							}
						}
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/FailureResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Messages"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
		"/messages/{messageID}": {
			"get": {
				"operationId": "GetMessageByID",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Message"
								}
							}
						}
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/FailureResponse"
								}
							}
						}
					},
					"401": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/FailureResponse"
								}
							}
						}
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/FailureResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Messages"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "messageID",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/messages/send": {
			"post": {
				"operationId": "SendMessage",
				"responses": {
					"200": {
						"description": ""
					},
					"204": {
						"description": "No content"
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/FailureResponse"
								}
							}
						}
					},
					"401": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/FailureResponse"
								}
							}
						}
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/FailureResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Messages"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SendMessageRequest"
							}
						}
					}
				}
			}
		},
		"/messages/delete": {
			"post": {
				"operationId": "DeleteMessage",
				"responses": {
					"200": {
						"description": ""
					},
					"204": {
						"description": "No content"
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/FailureResponse"
								}
							}
						}
					},
					"401": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/FailureResponse"
								}
							}
						}
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/FailureResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Messages"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/DeleteMessageRequest"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "http://127.0.0.1:3000/v1"
		}
	]
}